linting: sort imports, flake8 config, fixtures to conftest
