+++ setup.cfg
@@ -54,6 +54,7 @@ addopts =
 # ignore = D203*/
 exclude = .git,__pycache__,docs/conf.py,docs/_build,dist,features
 max-complexity = 10
+max-line-length = 95
 
 [aliases]
 docs = build_sphinx
