+++ tests/test_model_inputs.py
@@ -2,9 +2,7 @@
 """
 import numpy as np
 from numpy.testing import assert_equal
-
 from smif.inputs import ModelInputs
-from . fixtures.water_supply import one_input, two_inputs, one_dependency
 
 
 class TestModelInputs:
